requirements_python.txt
# ARP tools
scapy-python3

# Telepot library for bot telegram
telepot
